services/eval-service/requirements.txt
# Django and REST framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Server
gunicorn==21.2.0

# Consul integration
python-consul==1.1.0
requests==2.31.0

# Metrics (Prometheus)
django-prometheus==2.3.1

# Utilities
python-dotenv==1.0.0

# JWT
PyJWT==2.8.0

# Filtering
django-filter==23.5
